refactor: cleanup code, foldering and imports
