--- components/about/Bio.jsx
@@ -1,6 +1,0 @@
-const AboutBio = () => {
-    return <div>I'm ben I'm soooooo coool and good at poetry</div>;
-  };
-  
-  export default AboutBio;
-
