--- pages/about/bio.js
@@ -1,15 +1,0 @@
-
-import Bio from '../../components/about/Bio'
-
-
-
-
-export default function BioPage() {
-  return (
-    <>
-     
-      
-      <Bio />
-    </>
-  )
-}
