+++ styles/global.css
@@ -5,7 +5,7 @@
 /* clean up fonts */
 
 * {
-  font-family: "Crimson Text", serif;
+  font-family: 'Crimson Text', serif;
 
   /* for PRANKING */
   /* font-family: 'Gloria Hallelujah', cursive; */
@@ -16,11 +16,7 @@
 body {
   margin: 0 auto;
   max-width: 1000px;
-  
 }
-
-
-
 
 main {
   margin: 0 2rem;
@@ -35,15 +31,13 @@ li {
 }
 
 .nameHeader {
-  font-family: "Crimson Pro", serif;
+  font-family: 'Crimson Pro', serif;
 }
 
 .grey-text {
-  color: #000000ba
+  color: #000000ba;
 }
 
 .title-line-height {
-  line-height: .6;
+  line-height: 0.6;
 }
-
-
